Got the BSP Tree constructor to work
Now after the setup phase, the tree will have all the information necessary to create a renderPriorityList, I just haven't gotten to that yet. I tried ti out with multiple LineSegs and it looks like it works pretty well. One problem I noticed is that it once created a LineSeg with zero length (point 0 and 1 were the same). This isn't a huge problem, but my hunch is it comes from edge cases in the inFrontOf() method. I'll check on it later since that doesn't have a huge effect.